+++ doc/src/release-notes.md
@@ -4,6 +4,8 @@
 
 *Added:*
 
+* Add 'extended' as a partition for Frontier (#389).
+
 *Changed:*
 
 * Build the documentation with mdBook 0.5.4 (#360).